Further repackaging of Bayesian games and analytics
