+++ modules/engine/package.yaml
@@ -12,9 +12,7 @@ extra-source-files:
 library:
   source-dirs: src
   exposed-modules:
-    - OpenGames.Engine.BayesianGames
-    - OpenGames.Engine.Diagnostics
-    - OpenGames.Engine.Engine
+    - OpenGames.Engine.OpenGames.BayesianGames
     - OpenGames.Engine.OpenGames
     - OpenGames.Engine.Optics
     - OpenGames.Engine.Optics.StochasticStateful
